fix: refactor dropdown
