+++ lib/src/features/filter_workout/logic/filter_workout_bloc.dart
@@ -18,7 +18,6 @@ class FilterWorkoutBloc extends Cubit<FilterWorkoutState> {
   }
 
   void apply() {
-    print(state);
     emit(state.copyWith(
       disciplineActivity: state.disciplineActivityOnFilter,
       entryFee: state.entryFeeOnFilter,
